extract exercise cards into own component file
